# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pixel WEB SERVER",
    "version": "1.0.0",
    "description": "Документація до беку"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "https",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/admin/sections": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Створити нову секцію",
        "description": "Цей ендпоінт дозволяє створити нову секцію та зберегти її в базі даних. Секція може містити підсекції та продукти.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "photo": {
                    "type": "string",
                    "description": "URL фотографії секції",
                    "example": "https://example.com/photo.jpg"
                  },
                  "name": {
                    "type": "string",
                    "description": "Назва секції",
                    "example": "Секція 1"
                  },
                  "subSections": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "description": "ID підсекцій, що належать до цієї секції",
                      "example": "603c6b2f7f2e1b001f64792b"
                    },
                    "description": "Список ID підсекцій"
                  },
                  "products": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "description": "ID продуктів, що належать до цієї секції",
                      "example": "603c6b2f7f2e1b001f64792a"
                    },
                    "description": "Список ID продуктів"
                  }
                },
                "required": [
                  "name"
                ]
              },
              "example": {
                "photo": "https://example.com/photo.jpg",
                "name": "Секція 1",
                "subSections": [
                  "603c6b2f7f2e1b001f64792b"
                ],
                "products": [
                  "603c6b2f7f2e1b001f64792a"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Секція успішно створена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Повідомлення про успішне створення секції",
                      "example": "Successfully created"
                    },
                    "section": {
                      "type": "object",
                      "description": "Дані створеної секції",
                      "properties": {
                        "photo": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "subSections": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "products": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "example": {
                        "photo": "https://example.com/photo.jpg",
                        "name": "Секція 1",
                        "subSections": [
                          "603c6b2f7f2e1b001f64792b"
                        ],
                        "products": [
                          "603c6b2f7f2e1b001f64792a"
                        ],
                        "createdAt": "2024-09-11T00:00:00Z"
                      }
                    }
                  },
                  "required": [
                    "message",
                    "section"
                  ]
                },
                "example": {
                  "message": "Successfully created",
                  "section": {
                    "photo": "https://example.com/photo.jpg",
                    "name": "Секція 1",
                    "subSections": [
                      "603c6b2f7f2e1b001f64792b"
                    ],
                    "products": [
                      "603c6b2f7f2e1b001f64792a"
                    ],
                    "createdAt": "2024-09-11T00:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Помилка запиту через неправильний формат даних або відсутність обов'язкових полів",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Опис помилки",
                      "example": "Invalid data format"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Invalid data format"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Опис помилки",
                      "example": "Server error"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Server error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Видалення секцій",
        "description": "Цей ендпоінт дозволяє видаляти секції з бази даних. Можна видалити всі секції або лише вибрані, вказавши їх ID.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "description": "ID секцій для видалення. Якщо не вказано, видаляються всі секції.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "example": [
                "603c6b2f7f2e1b001f64792c"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Секції успішно видалені",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Повідомлення про успішне видалення.",
                      "example": "Deleted"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": "Deleted"
                }
              }
            }
          },
          "400": {
            "description": "Помилка запиту через некоректні дані",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Опис помилки",
                      "example": "Invalid request data"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Invalid request data"
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Опис помилки",
                      "example": "Server error"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Server error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}